requirements.txt
# CT Perfusion Analysis Dependencies (최소 버전)
# 필수 패키지
PyQt5>=5.15.0          # GUI 프레임워크
pydicom>=2.3.0         # DICOM 파일 읽기
numpy>=1.21.0,<2.0     # 수치 계산 (matplotlib 호환성)
Pillow>=9.0.0          # 이미지 변환 (PNG 생성)
scipy>=1.7.0           # 윤곽선 검출 (ndimage)

# NIfTI 파일 저장용 (웹 뷰어에서 사용)
nibabel>=3.2.0         # NIfTI 파일 저장

# 미사용 (제거 가능)
# pandas>=1.3.0        # 현재 코드에서 미사용
# matplotlib>=3.5.0    # 2D/3D 렌더링 제거로 불필요
